pyproject.toml
[tool.poetry]
name = "thermondo-task"
version = "0.1.0"
description = ""
authors = ["OM <om@bornig.com>"]

[tool.poetry.dependencies]
python = "^3.10.4"
Django = "^4.1"
bcrypt = "^3.2.2"
django-filter = "^22.1"
psycopg2-binary = "^2.9.3"
django-admin-autocomplete-filter = "^0.7.1"
django-cors-headers = "^3.13.0"
python-dotenv = "^0.20.0"
pycurl = "7.44.1"
uWSGI = "^2.0.20"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"
pytest-django = "^4.5.2"
django-taggit = "^3.0.0"

[tool.poetry.dev-dependencies]
coverage = "^5.5"
mypy = "^0.971"
black = "^22.6.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.10.1"
flake8-isort = "^5.0.0"
flake8 = "^5.0.4"
flake8-comprehensions = "^3.10.0"
django-stubs = "^1.12.0"

[build-system]
requires = ["poetry>=1.1.0"]
build-backend = "poetry.masonry.api"

[tool.django-stubs]
django_settings_module = "settings.base"

[tool.black]
line-length = 120
exclude = '''
(
  /(
    | \.build
    | \.venv
  )/
)
'''
